Now shows confirmation dialog for sketch deletion

+++ src/components/Sketch.vue
@@ -6,7 +6,7 @@
     <v-card-text v-html="body"></v-card-text>
     <v-card-actions>
       <v-spacer></v-spacer>
-      <v-btn v-if="sketch.createdBy && sketch.createdBy.uid && $globals.currentUser.uid == sketch.createdBy.uid" flat icon color="error" title="Delete" @click="deleteSketch(sketch)">
+      <v-btn v-if="sketch.createdBy && sketch.createdBy.uid && $globals.currentUser.uid == sketch.createdBy.uid" flat icon color="error" title="Delete this sketch" @click="deleteSketch(sketch)">
         <v-icon>delete</v-icon>
       </v-btn>
     </v-card-actions>
@@ -17,8 +17,37 @@
 import Firebase from "firebase";
 import EventBus from "@/service/EventBus.js";
 import { db } from "@/firebase";
+
 const sketches = db.collection("sketches");
 const deletedSketches = db.collection("deletedSketches");
+
+const animalNames = [
+  "Bearded Dragon",
+  "Sugar Glider",
+  "Hedgehog",
+  "Fennec Fox",
+  "Kinkajou",
+  "Serval",
+  "Wallaby",
+  "Capybara",
+  "Squirrel Monkey",
+  "Slow Loris",
+  "Giraffe",
+  "Chameleon",
+  "Duckling",
+  "Deer",
+  "Hippo",
+  "Kitten",
+  "Dolphin",
+  "Panda",
+  "Lamb",
+  "Seal",
+  "Puffer Fish",
+  "Bunny",
+  "Piglet",
+  "Hamster",
+  "Penguin"
+];
 
 export default {
   props: ["sketch"],
@@ -31,21 +60,32 @@ export default {
     }
   },
   methods: {
-    deleteSketch: sketch => {
-      sketch.deleted = Firebase.firestore.FieldValue.serverTimestamp();
-      deletedSketches
-        .add(sketch)
-        .then(() => {
-          sketches.doc(sketch.id).delete();
-        })
-        .then(() => {
-          EventBus.info(`Sketch '${sketch.title}' removed.`);
-        })
-        .catch(error => {
-          // eslint-disable-next-line
-          console.error(`Could not remove sketch '${sketch.title}'.`, error);
-          EventBus.info(`Could not remove sketch '${sketch.title}'.`);
-        });
+    deleteSketch(sketch) {
+      const animal =
+        animalNames[Math.floor(Math.random() * animalNames.length)];
+      this.$confirm(
+        `Every time you delete a sketch a ${animal} dies <span style="white-space: nowrap">:(</span><br>Are you sure you want to do this?`,
+        {
+          title: "Delete sketch?"
+        }
+      ).then(yes => {
+        if (yes) {
+          sketch.deleted = Firebase.firestore.FieldValue.serverTimestamp();
+          deletedSketches
+            .add(sketch)
+            .then(() => {
+              sketches.doc(sketch.id).delete();
+            })
+            .then(() => {
+              EventBus.info(`Sketch '${sketch.title}' removed.`);
+            })
+            .catch(error => {
+              // eslint-disable-next-line
+              console.error(`Could not remove sketch '${sketch.title}'.`, error);
+              EventBus.info(`Could not remove sketch '${sketch.title}'.`);
+            });
+        }
+      });
     }
   }
 };
